+++ changelog.md
@@ -5,6 +5,14 @@ This change log records updates after 2025-3-24.
 
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
+## [v0.6.2] - 2025-5-27
+### Added
+- Add tests for `MaxPooling1D` layer
+- Add more doc comments for `MaxPooling1D` layer
+
+### Changed
+- Introduced a shared `compute_output_shape` function to streamline output shape calculations across pooling layers.
+
 ## [v0.6.2] - 2025-5-26
 ### Added
 - Add `MaxPooling1D` layer implementation